fix nits

+++ cli/cmd/repair.go
@@ -87,22 +87,19 @@ func repair(ctx context.Context, forced bool) error {
 		return err
 	}
 
-	if !forced {
-		if serverVersion != clientVersion {
-
-			// Suggest directly upgrading to 2.9.4 or above for older versions
-			if repairNotApplicableVersionRegex.Match([]byte(serverVersion)) {
-				return fmt.Errorf("repair command is only applicable to 2.9 control-plane versions. Please try upgrading to the latest supported versions of Linkerd i.e 2.9.4 and above")
-			}
-
-			// Suggest 2.9.4 CLI version for all 2.9 server versions
-			if repairApplicableVersionRegex.Match([]byte(serverVersion)) {
-				return fmt.Errorf("Please run the repair command with a `stable-2.9.4` CLI.\nRun `export LINKERD2_VERSION=\"stable-2.9.4\"; curl -sL https://run.linkerd.io/install | sh; unset LINKERD2_VERSION` to install the server version of the CLI")
-			}
-
-			// Suggest server version for everything else. This includes all edge versions
-			return fmt.Errorf("Please run the repair command with a CLI that has the same version as the control plane.\nRun `export LINKERD2_VERSION=\"%s\"; curl -sL https://run.linkerd.io/install | sh; unset LINKERD2_VERSION` to install the server version of the CLI", serverVersion)
+	if !forced && serverVersion != clientVersion {
+		// Suggest directly upgrading to 2.9.4 or above for older versions
+		if repairNotApplicableVersionRegex.Match([]byte(serverVersion)) {
+			return fmt.Errorf("repair command is only applicable to 2.9 control-plane versions. Please try upgrading to the latest supported versions of Linkerd i.e 2.9.4 and above")
 		}
+
+		// Suggest 2.9.4 CLI version for all 2.9 server versions
+		if repairApplicableVersionRegex.Match([]byte(serverVersion)) {
+			return fmt.Errorf("Please run the repair command with a `stable-2.9.4` CLI.\nRun `export LINKERD2_VERSION=\"stable-2.9.4\"; curl -sL https://run.linkerd.io/install | sh; unset LINKERD2_VERSION` to install the server version of the CLI")
+		}
+
+		// Suggest server version for everything else. This includes all edge versions
+		return fmt.Errorf("Please run the repair command with a CLI that has the same version as the control plane.\nRun `export LINKERD2_VERSION=\"%s\"; curl -sL https://run.linkerd.io/install | sh; unset LINKERD2_VERSION` to install the server version of the CLI", serverVersion)
 	}
 
 	// Load the stored config
@@ -123,7 +120,6 @@ func repair(ctx context.Context, forced bool) error {
 		return fmt.Errorf("Failed to load values from linkerd-config: %s", err)
 	}
 
-	// Reset version fields
 	err = resetVersion(&values)
 	if err != nil {
 		return fmt.Errorf("Failed to reset version fields in linkerd-config: %s", err)
@@ -178,22 +174,16 @@ func resetVersion(values *linkerd2.Values) error {
 		return err
 	}
 
-	if values.DebugContainer != nil {
-		if values.DebugContainer.Image != nil {
-			values.DebugContainer.Image.Version = defaults.DebugContainer.Image.Version
-		}
+	if values.DebugContainer != nil && values.DebugContainer.Image != nil {
+		values.DebugContainer.Image.Version = defaults.DebugContainer.Image.Version
 	}
 
-	if values.Proxy != nil {
-		if values.Proxy.Image != nil {
-			values.Proxy.Image.Version = defaults.Proxy.Image.Version
-		}
+	if values.Proxy != nil && values.Proxy.Image != nil {
+		values.Proxy.Image.Version = defaults.Proxy.Image.Version
 	}
 
-	if values.ProxyInit != nil {
-		if values.ProxyInit.Image != nil {
-			values.ProxyInit.Image.Version = defaults.ProxyInit.Image.Version
-		}
+	if values.ProxyInit != nil && values.ProxyInit.Image != nil {
+		values.ProxyInit.Image.Version = defaults.ProxyInit.Image.Version
 	}
 
 	values.CliVersion = defaults.CliVersion
